Adding WPS to the process started by pressing button B on start up.
To test we suggest to delete the hole flash of Kniwwelino with 
esptool.py -p /dev/ttyUSB0 erase_flash 

Then you can compile any sketch against the new library version 1.3.3

Press the WPS button on you WiFi reouter or repeater and reboot the 
Kniwwelino with button b pressed. You should see now a WiFi like icon on 
the Kniwwelino matrice. In case of a failure  the Wifi-AP Message will 
apear and you can connect to the WiFiManager as usual. In the other case 
the connection to your WiFi should be established and your program 
should start.