Refine how the WPS will be accessible.
To use the WPS function. You reboot your Kniwwelino with button B 
pressed. When it rebooted and is showing the Wifi-AP message you can 
interupt the WiFiManager with button A. And after about a second a Wifi 
icon is blinking on the matrix. After the process the Kniwwelino should 
reconnect (you will see it on the matrice) or it will show No-WiFi.